Update! Fix the world-gen getter!

+++ src/main/kotlin/club/someoneice/oto/data/gem/helper/OreHelper.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 data class OreHelper(
     var ore_name: String,
-    val ore_distrib: HashMap<Int, Int>,
+    val ore_distrib: HashMap<Int, Double>,
     val silktouch: Boolean,
     val dropsList: ArrayList<DropHelper>,
     var dim: String
